backend/ocr_service/requirements.txt
flask
pillow
paddleocr
paddlepaddle
gunicorn